+++ osu.Game.Rulesets.Bms.Tests/BmsSkinTransformerTest.cs
@@ -411,29 +411,31 @@ namespace osu.Game.Rulesets.Bms.Tests
         }
 
         [Test]
-        public void TestAsyncHostSupportsOrdinaryNoteAndLongNoteHead()
+        public void TestAsyncHostSupportsOrdinaryNoteAndLongNoteCaps()
         {
             var noteHost = new BmsAsyncNoteDrawable(new BmsNoteSkinLookup(BmsNoteSkinElements.Note, 1, false, BmsKeymode.Key7K));
             var headHost = new BmsAsyncNoteDrawable(new BmsNoteSkinLookup(BmsNoteSkinElements.LongNoteHead, 1, false, BmsKeymode.Key7K));
+            var tailHost = new BmsAsyncNoteDrawable(new BmsNoteSkinLookup(BmsNoteSkinElements.LongNoteTail, 1, false, BmsKeymode.Key7K));
 
             Assert.Multiple(() =>
             {
                 Assert.That(noteHost.Drawable, Is.TypeOf<DefaultBmsNoteDisplay>());
                 Assert.That(headHost.Drawable, Is.TypeOf<DefaultBmsLongNoteHeadDisplay>());
+                Assert.That(tailHost.Drawable, Is.TypeOf<DefaultBmsLongNoteTailDisplay>());
+                Assert.That(tailHost.Drawable!.Alpha, Is.Zero);
             });
         }
 
-        [TestCase(BmsNoteSkinElements.LongNoteBody)]
-        [TestCase(BmsNoteSkinElements.LongNoteTail)]
-        public void TestAsyncHostRejectsNonCriticalLongNoteElements(BmsNoteSkinElements element)
+        [Test]
+        public void TestAsyncHostRejectsLongNoteBody()
         {
             Assert.That(
-                () => new BmsAsyncNoteDrawable(new BmsNoteSkinLookup(element, 1, false, BmsKeymode.Key7K)),
+                () => new BmsAsyncNoteDrawable(new BmsNoteSkinLookup(BmsNoteSkinElements.LongNoteBody, 1, false, BmsKeymode.Key7K)),
                 Throws.ArgumentException);
         }
 
         [Test]
-        public void TestDrawableLongNoteHeadUsesAsyncHostWhileBodyAndTailStaySkinnable()
+        public void TestDrawableLongNoteCapsUseAsyncHostWhileBodyStaysSkinnable()
         {
             var head = new DrawableBmsHitObject(new BmsHoldNoteHead { LaneIndex = 1, Keymode = BmsKeymode.Key7K });
             var body = new DrawableBmsHitObject(new BmsHoldNote { LaneIndex = 1, Keymode = BmsKeymode.Key7K });
@@ -445,8 +447,9 @@ namespace osu.Game.Rulesets.Bms.Tests
                 Assert.That(head.ChildrenOfType<SkinnableDrawable>(), Is.Empty);
                 Assert.That(body.ChildrenOfType<BmsAsyncNoteDrawable>(), Is.Empty);
                 Assert.That(body.ChildrenOfType<SkinnableDrawable>().Count(), Is.EqualTo(1));
-                Assert.That(tail.ChildrenOfType<BmsAsyncNoteDrawable>(), Is.Empty);
-                Assert.That(tail.ChildrenOfType<SkinnableDrawable>().Count(), Is.EqualTo(1));
+                Assert.That(tail.ChildrenOfType<BmsAsyncNoteDrawable>().Count(), Is.EqualTo(1));
+                Assert.That(tail.ChildrenOfType<SkinnableDrawable>(), Is.Empty);
+                Assert.That(tail.Height, Is.EqualTo(22.5f));
             });
         }
 
@@ -609,6 +612,15 @@ namespace osu.Game.Rulesets.Bms.Tests
             var transformer = new BmsRuleset().CreateSkinTransformer(new TestSkin(noteComponent: noteComponent), new BmsBeatmap());
 
             Assert.That(transformer!.GetDrawableComponent(new BmsNoteSkinLookup(BmsNoteSkinElements.LongNoteBody, 0, false)), Is.SameAs(noteComponent));
+        }
+
+        [Test]
+        public void TestCustomLongNoteTailTakesPriorityOverTransformerFallback()
+        {
+            var noteComponent = new Container();
+            var transformer = new BmsRuleset().CreateSkinTransformer(new TestSkin(noteComponent: noteComponent), new BmsBeatmap());
+
+            Assert.That(transformer!.GetDrawableComponent(new BmsNoteSkinLookup(BmsNoteSkinElements.LongNoteTail, 0, false)), Is.SameAs(noteComponent));
         }
 
         [Test]