+++ osu.Game.Rulesets.Bms/UI/DrawableBmsHitObject.cs
@@ -214,7 +214,7 @@ namespace osu.Game.Rulesets.Bms.UI
         {
             if (createLookup(hitObject) is BmsNoteSkinLookup lookup)
             {
-                if (lookup.Element is BmsNoteSkinElements.Note or BmsNoteSkinElements.LongNoteHead)
+                if (lookup.Element is BmsNoteSkinElements.Note or BmsNoteSkinElements.LongNoteHead or BmsNoteSkinElements.LongNoteTail)
                 {
                     return new BmsAsyncNoteDrawable(lookup)
                     {